sylly.en-package {sylly.en}	R Documentation

The sylly.en Package

Description

Language Support for 'sylly' Package: English.

Details

Package: 	 sylly.en

Type: 	 Package

Version: 	 0.1-3

Date: 	 2018-03-19

Depends: 	 R (>= 3.1),sylly

Encoding: 	 UTF-8

License: 	 GPL (>= 3)

LazyLoad: 	 yes

URL: 	 http://reaktanz.de/?c=hacking&s=koRpus

Adds support for the English language to the 'sylly' package. Due to some restrictions on CRAN, the full package sources are only available from the project homepage. To ask for help, report bugs, suggest feature improvements, or discuss the global development of the package, please consider subscribing to the koRpus-dev mailing list (<http://korpusml.reaktanz.de>).

Author(s)

Meik Michalke

[Package sylly.en version 0.1-3 Index]