hyph.en {sylly.en}	R Documentation

Hyphenation patterns for English

Description

Hyphenation patterns for English to be used by hyphen. These data objects are not really intended to be used directly, but rather to be consulted by the hyphen() function without further user interaction.

Usage

hyph.en

Format

The pattern slot of each hyphenation pattern object has three colums:

orig

The original pattern in patgen style format.

char

Only the character elements of the pattern which can be matched to parts of an actual word.

nums

A code of digits defining the possibility to split syllables at respective places in this pattern.

Source

The patterns (as they are present in the "orig" column described above) were originally provided by the LaTeX developers[1], under the terms of the LaTeX Project Public License[2]. Refer to Liang (1983) for a detailed explaination. From these original patterns the values in the remaining columns were created using read.hyph.pat.

In case any changes to the patterns were necessary to be used in this package, they are documented in the ChangeLog for the sources package. The unchanged original patterns can be found under [1].

References

Liang, F.M. (1983). Word Hy-phen-a-tion by Com-put-er. Dissertation, Stanford University, Dept. of Computer Science.

[1] http://tug.ctan.org/tex-archive/language/hyph-utf8/tex/generic/hyph-utf8/patterns/

[2] http://www.ctan.org/tex-archive/macros/latex/base/lppl.txt

See Also

read.hyph.pat, manage.hyph.pat

Examples

library(sylly.en)
sampleText <- c("This", "is", "a", "rather", "stupid", "demonstration")
hyphen(sampleText, hyph.pattern="en")

[Package sylly.en version 0.1-3 Index]