hyph.support.en {sylly.en}	R Documentation

Language support for English

Description

This function adds support for English to the sylly package. You should not need to call it manually, as that is done automatically when this package is loaded.

Usage

hyph.support.en()

Details

In particular, a new set of hyphenation patterns is being added (see hyph.en). To use the patterns with hyphen, simply use the abbreviation:

"en" — English hyphenation patterns (UK)

"en.us" — English hyphenation patterns (US)

See Also

hyphen, set.hyph.support

[Package sylly.en version 0.1-3 Index]